+++ src/Sentry/Platforms/Android/BindableSentryOptions.cs
@@ -33,8 +33,8 @@ internal partial class BindableSentryOptions
         public bool? PrintUncaughtStackTrace { get; set; }
         public double? ProfilesSampleRate { get; set; }
         public TimeSpan? ReadTimeout { get; set; }
-        public bool? EnableAndroidSdkTracing { get; set; }
-        public bool? EnableAndroidSdkBeforeSend { get; set; }
+        public bool? EnableTracing { get; set; }
+        public bool? EnableBeforeSend { get; set; }
         public LogCatIntegrationType? LogCatIntegration { get; set; }
         public int? LogCatMaxLines { get; set; }
 
@@ -61,8 +61,8 @@ internal partial class BindableSentryOptions
             options.PrintUncaughtStackTrace = PrintUncaughtStackTrace ?? options.PrintUncaughtStackTrace;
             options.ProfilesSampleRate = ProfilesSampleRate ?? options.ProfilesSampleRate;
             options.ReadTimeout = ReadTimeout ?? options.ReadTimeout;
-            options.EnableAndroidSdkTracing = EnableAndroidSdkTracing ?? options.EnableAndroidSdkTracing;
-            options.EnableAndroidSdkBeforeSend = EnableAndroidSdkBeforeSend ?? options.EnableAndroidSdkBeforeSend;
+            options.EnableTracing = EnableTracing ?? options.EnableTracing;
+            options.EnableBeforeSend = EnableBeforeSend ?? options.EnableBeforeSend;
             options.LogCatIntegration = LogCatIntegration ?? options.LogCatIntegration;
             options.LogCatMaxLines = LogCatMaxLines ?? options.LogCatMaxLines;
         }