+++ src/Sentry/Platforms/Android/SentryOptions.cs
@@ -255,7 +255,7 @@ public partial class SentryOptions
         /// Gets or sets a value that indicates if tracing features are enabled on the embedded Android SDK.
         /// The default value is <c>false</c> (disabled).
         /// </summary>
-        public bool EnableAndroidSdkTracing { get; set; } = false;
+        public bool EnableTracing { get; set; } = false;
 
         /// <summary>
         /// Gets or sets a value that indicates if the <c>BeforeSend</c> callback set in <see cref="o:SetBeforeSend"/>
@@ -266,6 +266,6 @@ public partial class SentryOptions
         /// implement all of the same features that may be present in the event graph. Some optional elements may
         /// be stripped away during the round-tripping between the two SDKs.  Use with caution.
         /// </remarks>
-        public bool EnableAndroidSdkBeforeSend { get; set; } = false;
+        public bool EnableBeforeSend { get; set; } = false;
     }
 }